requirements.txt
numpy
scipy
scikit-learn
pillow
hSpy
tensorflow
keras
opencv-python
matplotlib
IPython
